chore(ledger): move sqlite tests to a separate file
The inline test module in sqlite.rs was extracted into its own file to reduce the size of the main implementation file and improve maintainability. The module is now loaded via a path attribute pointing to the new sqlite_tests.rs file.

Auto-committed-on: dragonfly
